+++ src/blend/image_g2d.c
@@ -40,7 +40,8 @@ ret_t TK_WEAK g2d_copy_image(bitmap_t* fb, bitmap_t* img, const rect_t* src, xy_
   return RET_NOT_IMPL;
 }
 
-ret_t TK_WEAK g2d_rotate_image(bitmap_t* fb, bitmap_t* img, const rect_t* src, lcd_orientation_t o) {
+ret_t TK_WEAK g2d_rotate_image(bitmap_t* fb, bitmap_t* img, const rect_t* src,
+                               lcd_orientation_t o) {
   (void)fb;
   (void)img;
   (void)src;
@@ -49,7 +50,7 @@ ret_t TK_WEAK g2d_rotate_image(bitmap_t* fb, bitmap_t* img, const rect_t* src, l
 }
 
 ret_t TK_WEAK g2d_blend_image(bitmap_t* fb, bitmap_t* img, const rect_t* dst, const rect_t* src,
-                      uint8_t global_alpha) {
+                              uint8_t global_alpha) {
   (void)fb;
   (void)img;
   (void)dst;
@@ -59,8 +60,8 @@ ret_t TK_WEAK g2d_blend_image(bitmap_t* fb, bitmap_t* img, const rect_t* dst, co
 }
 
 #ifdef WITH_FAST_LCD_PORTRAIT
-ret_t TK_WEAK g2d_rotate_image_ex(bitmap_t* dst, bitmap_t* src, const rect_t* src_r, xy_t dx, xy_t dy,
-                          lcd_orientation_t o) {
+ret_t TK_WEAK g2d_rotate_image_ex(bitmap_t* dst, bitmap_t* src, const rect_t* src_r, xy_t dx,
+                                  xy_t dy, lcd_orientation_t o) {
   (void)dst;
   (void)src;
   (void)src_r;
@@ -71,7 +72,7 @@ ret_t TK_WEAK g2d_rotate_image_ex(bitmap_t* dst, bitmap_t* src, const rect_t* sr
 }
 
 ret_t TK_WEAK g2d_blend_image_rotate(bitmap_t* dst, bitmap_t* src, const rectf_t* dst_r,
-                             const rectf_t* src_r, uint8_t alpha, lcd_orientation_t o) {
+                                     const rectf_t* src_r, uint8_t alpha, lcd_orientation_t o) {
   (void)dst;
   (void)src;
   (void)dst_r;