+++ src/tkc/async.c
@@ -83,7 +83,8 @@ ret_t async_call_init_ex(uint32_t max_threads, uint32_t min_threads) {
 
 ret_t async_call_init_ex2(uint32_t max_threads, uint32_t min_threads, uint32_t stack_size) {
   return_value_if_fail(s_async_thread_pool == NULL, RET_BAD_PARAMS);
-  s_async_thread_pool = action_thread_pool_create_ex(max_threads, min_threads, stack_size, TK_THREAD_PRIORITY_NORMAL);
+  s_async_thread_pool =
+      action_thread_pool_create_ex(max_threads, min_threads, stack_size, TK_THREAD_PRIORITY_NORMAL);
   return_value_if_fail(s_async_thread_pool != NULL, RET_BAD_PARAMS);
 
   return RET_OK;