+++ tests/theme_test.cc
@@ -3,8 +3,8 @@
 #include "base/theme.h"
 #include "base/widget.h"
 #include "tkc/buffer.h"
-#include "tools/theme_gen/theme_gen.h"
 #include "base/style_factory.h"
+#include "base/theme_xml.h"
 #include "gtest/gtest.h"
 #include <stdlib.h>
 
@@ -53,29 +53,34 @@ static const char* widget_types[] = {WIDGET_TYPE_WINDOW_MANAGER,
                                      WIDGET_TYPE_LIST_VIEW_H,
                                      WIDGET_TYPE_LIST_ITEM,
                                      NULL};
-void GenThemeData(uint8_t* buff, uint32_t size, uint32_t state_nr, uint32_t name_nr) {
-  ThemeGen g;
+theme_t* GenThemeData(uint32_t state_nr, uint32_t name_nr) {
+  str_t str;
+  str_init(&str, 100000);
+
   for (int32_t i = 0; widget_types[i]; i++) {
     const char* type = widget_types[i];
+    str_append_format(&str, 100, "<%s>\n", type);
+    str_append_format(&str, 100, "<style name=\"%s\">\n", TK_DEFAULT_STYLE);
     for (uint32_t state = 0; state < state_nr; state++) {
-      Style s(type, TK_DEFAULT_STYLE, state_names[state]);
+      str_append_format(&str, 100, "<%s ", state_names[state]);
       for (uint32_t k = 0; k < name_nr; k++) {
         char name[32];
         char value[32];
 
         snprintf(name, sizeof(name), "%d", k);
         snprintf(value, sizeof(value), "%d", k);
-
-        s.AddValue(name, k);
-        s.AddValue(name, value);
+        str_append_format(&str, 100, " %s=\"%s\"", name, value);
       }
-      g.AddStyle(s);
+      str_append_format(&str, 100, ">\n</%s>\n", state_names[state]);
     }
+    str_append_format(&str, 100, "</%s>\n", type);
+    str_append_format(&str, 100, "</style>");
   }
-  wbuffer_t wbuffer;
-  wbuffer_t* b = wbuffer_init(&wbuffer, buff, size);
-  g.Output(b);
-  wbuffer_deinit(b);
+
+  theme_t* t = theme_xml_create(str.str);
+  str_reset(&str);
+
+  return t;
 }
 
 TEST(Theme, saveLoad) {
@@ -113,20 +118,17 @@ TEST(Theme, basic) {
   uint8_t buff[40 * 10240];
   uint32_t state_nr = 5;
   uint32_t name_nr = 5;
-  theme_t t;
   const uint8_t* style_data;
 
-  memset(&t, 0x00, sizeof(t));
-  GenThemeData(buff, sizeof(buff), state_nr, name_nr);
-  t.data = buff;
+  theme_t* t = GenThemeData(state_nr, name_nr);
 
   for (int32_t i = 0; widget_types[i]; i++) {
     const char* type = widget_types[i];
     for (uint32_t state = 0; state < state_nr; state++) {
-      style_data = theme_find_style(&t, type, 0, state_names[state]);
+      style_data = theme_find_style(t, type, 0, state_names[state]);
       ASSERT_EQ(style_data != NULL, true);
-      ASSERT_EQ(tk_str_eq(theme_get_style_type(&t), THEME_DEFAULT_STYLE_TYPE), true);
-      style_t* s = style_factory_create_style(NULL, theme_get_style_type(&t));
+      ASSERT_EQ(tk_str_eq(theme_get_style_type(t), THEME_DEFAULT_STYLE_TYPE), true);
+      style_t* s = style_factory_create_style(NULL, theme_get_style_type(t));
       ASSERT_EQ(s != NULL, true);
       ASSERT_EQ(style_set_style_data(s, style_data, NULL), RET_OK);
       for (uint32_t k = 0; k < name_nr; k++) {
@@ -138,4 +140,6 @@ TEST(Theme, basic) {
       ASSERT_EQ(style_destroy(s), RET_OK);
     }
   }
+
+  theme_destroy(t);
 }